transforma página de pomodoro em componente de metas
