+++ front/goals-notes/src/App.js
@@ -9,7 +9,6 @@ import { Route,Routes, BrowserRouter } from "react-router-dom";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { verify } from "./utils/token_verify";
 import React, { useEffect } from 'react';
-import PomodoroTimer from "./pages/pomodoro/pomodoro";
 
   function App() {
     useEffect(() => {
@@ -31,7 +30,6 @@ import PomodoroTimer from "./pages/pomodoro/pomodoro";
           <Route path="/:id/UserUpdate" element={<UserUpdate />}></Route>
           <Route path="/:id/Amigos" element={<Amigos />}></Route>
           <Route path="/:id/Perfil" element={<Perfil />}></Route>
-          <Route path="/:id/Pomodoro" element={<PomodoroTimer/>}></Route>
         </Routes>
     </BrowserRouter>
   );
